+++ billy-france/src/test/java/com/premiumminds/billy/france/test/services/documents/handler/TestFRInvoiceIssuingHandler.java
@@ -19,6 +19,7 @@
 package com.premiumminds.billy.france.test.services.documents.handler;
 
 import com.premiumminds.billy.core.exceptions.SeriesUniqueCodeNotFilled;
+import com.premiumminds.billy.core.services.exceptions.DocumentSeriesDoesNotExistException;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -46,16 +47,19 @@ public class TestFRInvoiceIssuingHandler extends FRDocumentAbstractTest {
         try {
             FRInvoiceEntity invoice = this.newInvoice(INVOICE_TYPE.FT);
 
+            this.createSeries(invoice, this.DEFAULT_SERIES);
+
             this.issueNewInvoice(this.handler, invoice, this.DEFAULT_SERIES);
             this.issuedInvoiceUID = invoice.getUID();
-        } catch (DocumentIssuingException | SeriesUniqueCodeNotFilled e) {
+
+        } catch (DocumentIssuingException | DocumentSeriesDoesNotExistException | SeriesUniqueCodeNotFilled e) {
             e.printStackTrace();
         }
 
     }
 
     @Test
-    public void testIssuedInvoiceSimple() throws DocumentIssuingException {
+    public void testIssuedInvoiceSimple() {
         FRInvoice issuedInvoice = this.getInstance(DAOFRInvoice.class).get(this.issuedInvoiceUID);
 
         Assertions.assertEquals(this.DEFAULT_SERIES, issuedInvoice.getSeries());
@@ -65,7 +69,8 @@ public class TestFRInvoiceIssuingHandler extends FRDocumentAbstractTest {
     }
 
     @Test
-    public void testIssuedInvoiceSameSeries() throws DocumentIssuingException, SeriesUniqueCodeNotFilled {
+    public void testIssuedInvoiceSameSeries()
+            throws DocumentIssuingException, DocumentSeriesDoesNotExistException, SeriesUniqueCodeNotFilled {
         FRInvoice issuedInvoice = this.getInstance(DAOFRInvoice.class).get(this.issuedInvoiceUID);
         Integer nextNumber = 2;
 
@@ -85,13 +90,15 @@ public class TestFRInvoiceIssuingHandler extends FRDocumentAbstractTest {
     }
 
     @Test
-    public void testIssuedInvoiceDifferentSeries() throws DocumentIssuingException, SeriesUniqueCodeNotFilled {
+    public void testIssuedInvoiceDifferentSeries()
+            throws DocumentIssuingException, DocumentSeriesDoesNotExistException, SeriesUniqueCodeNotFilled {
         Integer nextNumber = 1;
         String newSeries = "FT NEW_SERIES";
 
         FRInvoiceEntity newInvoice = this.newInvoice(INVOICE_TYPE.FT);
 
         UID newInvoiceUID = newInvoice.getUID();
+        this.createSeries(newInvoice, newSeries);
 
         this.issueNewInvoice(this.handler, newInvoice, newSeries);
 
@@ -104,14 +111,24 @@ public class TestFRInvoiceIssuingHandler extends FRDocumentAbstractTest {
     }
 
     @Test
-    public void testIssuedInvoiceSameSourceBilling() throws DocumentIssuingException, SeriesUniqueCodeNotFilled {
+    public void testIssuedInvoiceSameSourceBilling()
+            throws DocumentIssuingException, DocumentSeriesDoesNotExistException, SeriesUniqueCodeNotFilled {
         FRInvoiceEntity newInvoice = this.newInvoice(INVOICE_TYPE.FT);
 
         UID newInvoiceUID = newInvoice.getUID();
+        this.createSeries(newInvoice, this.DEFAULT_SERIES);
 
         this.issueNewInvoice(this.handler, newInvoice, this.DEFAULT_SERIES);
 
         this.getInstance(DAOFRInvoice.class).get(newInvoiceUID);
+    }
+
+    @Test
+    public void testSeriesDoesNotExist() {
+        FRInvoiceEntity invoiceEntity = this.newInvoice(INVOICE_TYPE.FT);
+
+        Assertions.assertThrows(DocumentSeriesDoesNotExistException.class,
+                () -> this.issueNewInvoice(this.handler, invoiceEntity, "A RANDOM SERIES"));
     }
 
 }