+++ billy-france/src/test/java/com/premiumminds/billy/france/test/services/documents/handler/TestFRSimpleInvoiceIssuingHandlerWithTicket.java
@@ -60,7 +60,9 @@ public class TestFRSimpleInvoiceIssuingHandlerWithTicket extends FRDocumentAbstr
             this.setUpParamenters();
             this.parameters.setInvoiceSeries(this.DEFAULT_SERIES);
 
-            FRBusinessEntity business = new FRBusinessTestUtil(FRAbstractTest.injector).getBusinessEntity("business");
+			final String uid = "business";
+			this.createSeries(uid, this.DEFAULT_SERIES);
+			FRBusinessEntity business = new FRBusinessTestUtil(FRAbstractTest.injector).getBusinessEntity(uid);
             FRSimpleInvoice.Builder simpleInvoiceBuilder = new FRSimpleInvoiceTestUtil(FRAbstractTest.injector)
                     .getSimpleInvoiceBuilder(business, CLIENTTYPE.CUSTOMER);