+++ billy-portugal/src/main/java/com/premiumminds/billy/portugal/services/documents/PTCreditNoteIssuingHandler.java
@@ -19,6 +19,7 @@
 package com.premiumminds.billy.portugal.services.documents;
 
 import com.premiumminds.billy.core.exceptions.SeriesUniqueCodeNotFilled;
+import com.premiumminds.billy.core.services.exceptions.DocumentSeriesDoesNotExistException;
 import javax.inject.Inject;
 
 import com.premiumminds.billy.core.persistence.dao.DAOInvoiceSeries;
@@ -41,7 +42,7 @@ public class PTCreditNoteIssuingHandler extends PTGenericInvoiceIssuingHandler<P
 
     @Override
     public PTCreditNoteEntity issue(PTCreditNoteEntity document, PTIssuingParams parameters)
-            throws DocumentIssuingException, SeriesUniqueCodeNotFilled {
+            throws DocumentIssuingException, DocumentSeriesDoesNotExistException, SeriesUniqueCodeNotFilled {
 
         return this.issue(document, parameters, this.daoCreditNote, PTCreditNoteIssuingHandler.INVOICE_TYPE);
     }