+++ billy-portugal/src/main/java/com/premiumminds/billy/portugal/services/documents/PTReceiptInvoiceIssuingHandler.java
@@ -19,6 +19,7 @@
 package com.premiumminds.billy.portugal.services.documents;
 
 import com.premiumminds.billy.core.exceptions.SeriesUniqueCodeNotFilled;
+import com.premiumminds.billy.core.services.exceptions.DocumentSeriesDoesNotExistException;
 import javax.inject.Inject;
 
 import com.premiumminds.billy.core.persistence.dao.DAOInvoiceSeries;
@@ -42,7 +43,7 @@ public class PTReceiptInvoiceIssuingHandler
 
     @Override
     public PTReceiptInvoiceEntity issue(PTReceiptInvoiceEntity document, PTIssuingParams parameters)
-            throws DocumentIssuingException, SeriesUniqueCodeNotFilled {
+            throws DocumentIssuingException, DocumentSeriesDoesNotExistException, SeriesUniqueCodeNotFilled {
         return this.issue(document, parameters, this.daoReceiptInvoice, PTReceiptInvoiceIssuingHandler.INVOICE_TYPE);
     }