+++ billy-portugal/src/test/java/com/premiumminds/billy/portugal/test/services/documents/TestDocumentIssuingService.java
@@ -48,7 +48,6 @@ public class TestDocumentIssuingService extends PTDocumentAbstractTest {
 
     @Test
     public void testIssuingService() throws DocumentIssuingException {
-
         final PTBusinessEntity businessEntity = new PTBusinessTestUtil(PTAbstractTest.injector).getBusinessEntity();
         this.createSeries(businessEntity.getUID().toString(), "A");
         this.service.issue(