+++ billy-spain/src/test/java/com/premiumminds/billy/spain/test/services/documents/handler/TestESManualInvoiceIssuingHandler.java
@@ -19,6 +19,7 @@
 package com.premiumminds.billy.spain.test.services.documents.handler;
 
 import com.premiumminds.billy.core.exceptions.SeriesUniqueCodeNotFilled;
+import com.premiumminds.billy.core.services.exceptions.DocumentSeriesDoesNotExistException;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -45,15 +46,17 @@ public class TestESManualInvoiceIssuingHandler extends ESDocumentAbstractTest {
         try {
             ESInvoiceEntity invoice = this.newInvoice(INVOICE_TYPE.FT, SOURCE_BILLING.MANUAL);
 
+            this.createSeries(invoice, this.DEFAULT_SERIES);
+
             this.issueNewInvoice(this.handler, invoice, this.DEFAULT_SERIES);
             this.issuedInvoiceUID = invoice.getUID();
-        } catch (DocumentIssuingException | SeriesUniqueCodeNotFilled e) {
+        } catch (DocumentIssuingException | DocumentSeriesDoesNotExistException | SeriesUniqueCodeNotFilled e) {
             e.printStackTrace();
         }
     }
 
     @Test
-    public void testIssuedManualInvoiceSimple() throws DocumentIssuingException {
+    public void testIssuedManualInvoiceSimple() {
         ESInvoiceEntity issuedInvoice = this.getInstance(DAOESInvoice.class).get(this.issuedInvoiceUID);
 
         Assertions.assertEquals(this.DEFAULT_SERIES, issuedInvoice.getSeries());